requirements.txt
absl-py==0.7.1
astor==0.8.0
bokeh==1.2.0
certifi==2019.3.9
cycler==0.10.0
dj-database-url==0.5.0
Django==2.2.1
django-bootstrap4==0.0.8
django-heroku==0.3.1
gast==0.2.2
grpcio==1.21.1
gunicorn==19.9.0
h5py==2.9.0
Jinja2==2.10.1
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mock==3.0.5
numpy==1.16.4
packaging==19.0
Pillow==6.0.0
protobuf==3.7.1
psycopg2==2.8.2
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sqlparse==0.3.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tornado==6.0.2
Werkzeug==0.15.4
whitenoise==4.1.2
wordcloud==1.5.0
